+++ src/Auth/BasicCredentials.cs
@@ -3,7 +3,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 
-namespace Tavis
+namespace Tavis.Auth
 {
     public class BasicCredentials : HttpCredentials
     {