+++ ReleaseNotes.md
@@ -1,3 +1,6 @@
 # Tavis.Auth #
 
-##0.1.0
+##0.9.0
+- Added support for gateway keys that are always sent for the right URL space
+## 0.1.0 - beta2
+- First Release supporting credentials in response to www-auth headers